Created agents module splitting the agents
